fix(content API): Refactor default scenario to use multiple files.
The index file in a scenario should be used only to combine the scenario, for clarity
